これでそこそこ落ち着いたんじゃないかなぁ…

+++ lib/Devel/REPL/Plugin/HTTPDebug.pm
@@ -7,17 +7,21 @@ use namespace::clean -except => [ 'meta' ];
 use LWP::UserAgent;
 require HTTP::Request::Common;
 use Rose::URI;
-use Term::ANSIColor qw(:constants);
+use Term::ANSIColor;
+
+use Data::Dumper;
+$Data::Dumper::Indent = 1;
 
 my %query;
+my $format = 'raw';
 # cache for latest request and response
-my ($req, $res);
 
 our $VERSION = '0.01';
 
 our %COLORS = (
-    HEADER_KEY   => BOLD.BLUE,
-    HEADER_VALUE => ''
+    HEADER_KEY   => 'bold blue',
+    HEADER_VALUE => '',
+    CONTENT => ''
 );
 
 has ua => (
@@ -25,6 +29,16 @@ has ua => (
     isa => 'Object',
     lazy => 1,
     default => sub { LWP::UserAgent->new }
+);
+
+has req => (
+    is => 'rw',
+    isa => 'HTTP::Request',
+);
+
+has res => (
+    is => 'rw',
+    isa => 'HTTP::Response',
 );
 
 has cookie_file => (
@@ -42,12 +56,7 @@ has cookie => (
     is => 'rw',
     isa => 'HTTP::Cookies',
     lazy => 1,
-    default => sub {
-        my $self = shift;
-        my $file = $self->cookie_file;
-        require HTTP::Cookies;
-        return HTTP::Cookies->new(file => $file, autosave => 1);
-    }
+    default => sub { return shift->_new_cookie; }
 );
 
 has host => (
@@ -65,9 +74,21 @@ around 'eval' => sub {
     return $orig->(@_);
 };
 
+sub _new_cookie {
+    my $self = shift;
+    require HTTP::Cookies;
+    return HTTP::Cookies->new(file => $self->cookie_file, autosave => 1);
+}
+
+# 違う出力にしたいときは、aroundで変えてほしい
+sub _dumper {
+    require YAML;
+    return YAML->can('Dump');
+}
+
 sub add {
     my $self = shift;
-    my %p = ($_[1]) ? @_ : %{$_[0]};
+    my %p = %{$_[0]};
     while (my ($k, $v) = each %p) {
         $query{$k} = $v;
     }
@@ -101,28 +122,50 @@ sub clear_session {
 
 sub undef_session {
     my $self = shift;
-    $self->ua->cookie_jar(undef);
-    unlink $self->cookie_file;
-    my $c = HTTP::Cookies->new(file => $self->cookie_file, autosave => 1);
-    $self->cookie($c);
+    unlink $self->cookie_file if -e $self->cookie_file;
+    $self->cookie($self->_new_cookie);
 }
 
-sub hdump { coloring_line($req->as_string) }
+sub hdump {
+    my $self = shift;
+    coloring_line($self->req->as_string);
+}
 
 sub rdump {
-    my $h = $res->protocol . " ". $res->code . " " . $res->message;
+    my $self = shift;
+    my $res = $self->res;
+    my $h = join(' ', ($res->protocol, $res->code, $res->message));
     coloring_line($h . "\n" . $res->headers->as_string)
 }
 
-sub cdump { $res->decoded_content }
+sub cdump {
+    my $self = shift;
+    my $str = $self->res->decoded_content;
+    if ($format eq 'json') {
+        # ここも変えられたらいい気はするけど…
+        require JSON::XS;
+        my $dumped = _dumper->(JSON::XS::decode_json($str));
+        utf8::encode($dumped) if utf8::is_utf8($dumped);
+        return $dumped;
+    }
+    elsif ($format eq 'xml') {
+        # 本音は XML::Simple sucks
+        require XML::Simple;
+        utf8::encode($str) if utf8::is_utf8($str);
+        my $dumped = _dumper->(XMLin($str));
+        utf8::encode($dumped) if utf8::is_utf8($dumped);
+        return $dumped;
+    }
+    return $str;
+}
 
 sub coloring_line {
     my $line = shift;
     my @lines = split "\n", $line;
     @lines = map {
         if(my ($k, $v) = (/^(.+?) (.+?)$/)) {
-            $k = $COLORS{HEADER_KEY}   . $k . RESET if $COLORS{HEADER_KEY};
-            $v = $COLORS{HEADER_VALUE} . $v . RESET if $COLORS{HEADER_VALUE};
+            $k = colored($k, $COLORS{HEADER_KEY})   if $COLORS{HEADER_KEY};
+            $v = colored($v, $COLORS{HEADER_VALUE}) if $COLORS{HEADER_VALUE};
             "${k} ${v}";
         }
         else {
@@ -132,53 +175,52 @@ sub coloring_line {
     return join("\n", @lines)."\n";
 }
 
-sub get {
-    my $self = shift;
-    $res = $self->get_q(@_);
-    $self->print(rdump() ."\n". cdump());
+sub get  { shift->_req_common('get_q', @_);  }
+sub post { shift->_req_common('post_q', @_); }
+sub file { shift->_req_common('file_q', @_); }
+
+sub _req_common {
+    my ($self, $c, $path, $args, $res_format) = @_;
+    $self->$c($path, $args) or return;
+    $format = $res_format || 'raw';
+    $self->print($self->rdump ."\n". $self->res->decoded_content);
 }
 
 sub get_q {
     my $self = shift;
     my $u = $self->_get_url(@_) or return;
     $u->query_form(%query);
-    $req = HTTP::Request->new(GET => "$u");
-    $res = $self->ua->request($req);
-}
 
-sub post {
-    my $self = shift;
-    $res = $self->post_q(@_) or return;
-    $self->print(rdump() ."\n". cdump());
+    my $req = HTTP::Request->new(GET => "$u");
+    $self->req($req);
+    $self->res($self->ua->request($self->req));
 }
 
 sub post_q {
     my $self = shift;
     my $u = $self->_get_url(@_) or return;
-    $req = HTTP::Request::Common::POST("$u", [ %query ]);
-    $res = $self->ua->request($req);
-}
 
-sub file {
-    my $self = shift;
-    my $res = $self->file_q(@_) or return;
-    $self->print(rdump() ."\n". cdump());
+    my $req = HTTP::Request::Common::POST("$u", [ %query ]);
+    $self->req($req);
+    $self->res($self->ua->request($self->req));
 }
 
 sub file_q {
     my $self = shift;
     my $u = $self->_get_url(@_) or return;
-    $req = HTTP::Request::Common::POST(
+
+    my $req = HTTP::Request::Common::POST(
         "$u",
         Content_Type => 'form-data',
         Content => [ %query ]
     );
-    $res = $self->ua->request($req);
+    $self->req($req);
+    $self->res($self->ua->request($req));
 }
 
 sub _get_url {
     my $self = shift;
-    my ($path, @p) = @_;
+    my ($path, $p) = @_;
     if (!$self->host) {
         $self->print('no host');
         return;
@@ -187,7 +229,7 @@ sub _get_url {
     $u->scheme('http');
     $u->host($self->host);
     $u->path($path);
-    $self->add(@p) if @p;
+    $self->add($p) if %$p;
     return $u;
 }
 
@@ -200,7 +242,7 @@ Devel::REPL::Plugin::HTTPDebug -
 
 =head1 SYNOPSIS
 
-  use Devel::REPL::Plugin::HTTPDebug;
+  $ ./bin/http_debug.pl
 
 =head1 DESCRIPTION
